+++ crates/configuration/src/lib.rs
@@ -7,6 +7,7 @@ pub mod version3;
 
 pub use configuration::{
     introspect, parse_configuration, Configuration, RawConfiguration, CONFIGURATION_FILENAME,
+    CONFIGURATION_JSONSCHEMA_FILENAME,
 };
 pub use error::Error;
 pub use values::{ConnectionUri, IsolationLevel, PoolSettings, Secret};