+++ crates/tests/databases-tests/src/postgres/mutation_tests.rs
@@ -87,6 +87,26 @@ mod basic {
 
         insta::assert_json_snapshot!(result)
     }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn v1_insert_enum() {
+        let ndc_metadata =
+            FreshDeployment::create(common::CONNECTION_URI, common::CHINOOK_NDC_METADATA_PATH)
+                .await
+                .unwrap();
+
+        let router = tests_common::router::create_router(
+            &ndc_metadata.ndc_metadata_path,
+            &ndc_metadata.connection_uri,
+        )
+        .await;
+
+        let mutation_result = run_mutation(router.clone(), "v1_insert_enum").await;
+
+        let result = mutation_result;
+
+        insta::assert_json_snapshot!(result)
+    }
 }
 
 #[cfg(test)]
@@ -149,5 +169,30 @@ mod negative {
         let result = mutation_result;
 
         insta::assert_json_snapshot!(result);
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn v1_insert_enum_invalid_label() {
+        let ndc_metadata =
+            FreshDeployment::create(common::CONNECTION_URI, common::CHINOOK_NDC_METADATA_PATH)
+                .await
+                .unwrap();
+
+        let router = tests_common::router::create_router(
+            &ndc_metadata.ndc_metadata_path,
+            &ndc_metadata.connection_uri,
+        )
+        .await;
+
+        let mutation_result = run_mutation_fail(
+            router.clone(),
+            "v1_insert_enum_invalid_label",
+            StatusCode::UNPROCESSABLE_ENTITY,
+        )
+        .await;
+
+        let result = mutation_result;
+
+        insta::assert_json_snapshot!(result)
     }
 }